only delete upload files from the temp directory if the debug flag is set to false, otherwise leave them there to aid in debugging

+++ src/main/java/eu/openminted/workflow/service/WorkflowServiceImpl.java
@@ -265,7 +265,7 @@ public class WorkflowServiceImpl implements WorkflowService {
 									dataset = uploadFileToHistory(history.getId(), path.toFile());
 
 								} finally {
-									if (path != null)
+									if (path != null && debug)
 										Files.delete(path);
 								}
 							}
